Extract custom mutation
